[Chore] 메인 뷰컨이랑 페이지 뷰컨 분리 #11
